backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pyotp==2.9.0
python-decouple==3.8
pyotp==2.9.0
qrcode==7.4.2
